Create session.rs
