+++ pg_analytics/src/api/init.rs
@@ -1,6 +1,6 @@
 use pgrx::*;
 
-use crate::datafusion::context::DatafusionContext;
+use crate::datafusion::session::Session;
 
 extension_sql!(
     r#"
@@ -11,9 +11,9 @@ extension_sql!(
 #[pg_guard]
 #[no_mangle]
 pub extern "C" fn init() {
-    let _ =
-        DatafusionContext::init(DatafusionContext::catalog_oid().expect("Catalog OID not found"))
-            .unwrap_or_else(|err| {
-                panic!("{}", err);
-            });
+    let _ = Session::init(Session::catalog_oid().expect("Catalog OID not found")).unwrap_or_else(
+        |err| {
+            panic!("{}", err);
+        },
+    );
 }
