+++ pg_analytics/src/datafusion/mod.rs
@@ -5,6 +5,7 @@ pub mod context;
 pub mod datatype;
 pub mod directory;
 pub mod schema;
+pub mod session;
 pub mod stream;
 pub mod table;
 pub mod writer;
