+++ pg_analytics/src/hooks/truncate.rs
@@ -1,6 +1,6 @@
 use pgrx::*;
 
-use crate::datafusion::context::DatafusionContext;
+use crate::datafusion::session::Session;
 use crate::datafusion::table::DatafusionTable;
 use crate::errors::ParadeError;
 use crate::hooks::handler::IsColumn;
@@ -52,7 +52,7 @@ pub unsafe fn truncate(truncate_stmt: *mut pg_sys::TruncateStmt) -> Result<(), P
 
         pg_sys::RelationClose(relation);
 
-        DatafusionContext::with_tables(schema_name, |tables| async move {
+        Session::with_tables(schema_name, |tables| async move {
             let mut lock = tables.lock().await;
             let (mut delta_table, _) = lock.delete(&table_path, None).await?;
